+++ src/oemof/network/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.5.0a2"
+__version__ = "0.5.0a6"
 
 from . import energy_system
 from . import graph
@@ -7,6 +7,8 @@ from . import network
 from .network import Bus
 from .network import Component
 from .network import Edge
+from .network import Entity
+from .network import Node
 from .network import Sink
 from .network import Source
 from .network import Transformer
@@ -15,10 +17,12 @@ __all__ = [
     "Bus",
     "Component",
     "Edge",
+    "Entity",
     "energy_system",
     "graph",
     "groupings",
     "network",
+    "Node",
     "Sink",
     "Source",
     "Transformer",