+++ src/oemof/network/network/__init__.py
@@ -1,6 +1,8 @@
 from .edge import Edge
+from .entity import Entity
 from .nodes import Bus
 from .nodes import Component
+from .nodes import Node
 from .nodes import Sink
 from .nodes import Source
 from .nodes import Transformer
@@ -9,6 +11,8 @@ __all__ = [
     "Bus",
     "Component",
     "Edge",
+    "Entity",
+    "Node",
     "Sink",
     "Source",
     "Transformer",